some rework
