+++ src/gameData.ts
@@ -1,5 +1,3 @@
-import { ScoreTableItem } from "./entities/ScoreTableItem.type";
-
 export let games = new Map();
 export let players = new Map();
 export let rooms = new Map();
